Support more solar arrays

+++ src/battery.rs
@@ -3,5 +3,5 @@ pub struct Battery {
     pub capacity: f32,
     pub temperature_high: f32,
     pub temperature_low: f32,
-    pub voltage: f32,
+    pub voltage: f64,
 }
